requirements.txt
certifi==2024.7.4
elastic-transport==8.15.0
elasticsearch==8.15.0
numpy==2.0.1
pandas==2.2.2
psycopg2==2.9.9
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
six==1.16.0
tzdata==2024.1
urllib3==2.2.2
